Rename layout component
